Move yearRandomNumbers to random.h/random.cpp (#3162)
