+++ src/solver/simulation/include/antares/solver/simulation/random.h
@@ -21,12 +21,44 @@
 
 #pragma once
 
+#include <map>
+#include <vector>
+
 #include <antares/mersenne-twister/mersenne-twister.h>
-#include <antares/solver/simulation/solver_utils.h>
 #include <antares/study/study.h>
 
 namespace Antares::Solver::Simulation
 {
+class yearRandomNumbers
+{
+public:
+    yearRandomNumbers();
+    ~yearRandomNumbers() = default;
+
+    void setNbAreas(uint nbAreas);
+    void setPowerFluctuations(Data::PowerFluctuations powerFluctuations);
+    void reset();
+
+    // General data
+    uint pNbAreas;
+    Data::PowerFluctuations pPowerFluctuations;
+
+    // Data for thermal noises
+    std::vector<std::vector<double>> pThermalNoisesByArea;
+    std::vector<size_t> pNbClustersByArea;
+
+    // Data for reservoir levels
+    std::vector<double> pReservoirLevels;
+
+    // Data for unsupplied and spilled energy costs
+    std::vector<double> pUnsuppliedEnergy;
+    std::vector<double> pSpilledEnergy;
+
+    // Hydro costs noises
+    std::vector<std::vector<double>> pHydroCostsByArea_freeMod;
+    std::vector<double> pHydroCosts_rampingOrExcursion;
+};
+
 class randomNumbers
 {
 public:
