+++ src/solver/simulation/include/antares/solver/simulation/solver_utils.h
@@ -102,122 +102,6 @@ private:
     void writeUpdateTimes(IResultWriter& writer) const;
     void writeOptimizationTimeToOutput(IResultWriter& writer) const;
 };
-
-class yearRandomNumbers
-{
-public:
-    yearRandomNumbers()
-    {
-        pNbAreas = 0;
-    }
-
-    ~yearRandomNumbers() = default;
-
-    void setNbAreas(uint nbAreas)
-    {
-        pNbAreas = nbAreas;
-    }
-
-    void setPowerFluctuations(Data::PowerFluctuations powerFluctuations)
-    {
-        pPowerFluctuations = powerFluctuations;
-    }
-
-    void reset()
-    {
-        // Thermal noises
-        for (uint a = 0; a != pNbAreas; a++)
-        {
-            pThermalNoisesByArea[a].assign(pThermalNoisesByArea[a].size(), 0);
-        }
-
-        // General
-        pNbClustersByArea.assign(pNbAreas, 0);
-
-        // Reservoir levels, spilled and unsupplied energy costs
-        pReservoirLevels.assign(pNbAreas, 0);
-        pUnsuppliedEnergy.assign(pNbAreas, 0);
-        pSpilledEnergy.assign(pNbAreas, 0);
-
-        // Hydro costs noises
-        switch (pPowerFluctuations)
-        {
-        case Data::lssFreeModulations:
-        {
-            for (uint a = 0; a != pNbAreas; a++)
-            {
-                pHydroCostsByArea_freeMod[a].assign(8784, 0);
-            }
-            break;
-        }
-
-        case Data::lssMinimizeRamping:
-        case Data::lssMinimizeExcursions:
-        {
-            pHydroCosts_rampingOrExcursion.assign(pNbAreas, 0);
-            break;
-        }
-
-        case Data::lssUnknown:
-            break;
-        }
-    }
-
-    // General data
-    uint pNbAreas;
-    Data::PowerFluctuations pPowerFluctuations;
-
-    // Data for thermal noises
-    std::vector<std::vector<double>> pThermalNoisesByArea;
-    std::vector<size_t> pNbClustersByArea;
-
-    // Data for reservoir levels
-    std::vector<double> pReservoirLevels;
-
-    // Data for unsupplied and spilled energy costs
-    std::vector<double> pUnsuppliedEnergy;
-    std::vector<double> pSpilledEnergy;
-
-    // Hydro costs noises
-    std::vector<std::vector<double>> pHydroCostsByArea_freeMod;
-    std::vector<double> pHydroCosts_rampingOrExcursion;
-};
-
-// Class representing a hydro cost noise.
-// This class allows sorting hydro costs noises into increasing absolute values order
-// when instances are contained in a :
-//		std::set<hydroCostNoise, compareHydroCostsNoises>
-class hydroCostNoise
-{
-public:
-    hydroCostNoise(double v, uint i):
-        value(v),
-        index(i)
-    {
-    }
-
-    inline double getValue() const
-    {
-        return value;
-    }
-
-    inline uint getIndex() const
-    {
-        return index;
-    }
-
-    double value;
-    uint index;
-};
-
-// Used to sort the hydro costs noises into increasing absolute values order in a std::set
-struct compareHydroCostsNoises
-{
-    inline bool operator()(const hydroCostNoise& hcnr1, const hydroCostNoise& hcnr2) const
-    {
-        return std::abs(hcnr1.getValue()) < std::abs(hcnr2.getValue());
-    }
-};
 } // namespace Antares::Solver::Simulation
 
 #endif // __SOLVER_SIMULATION_SOLVER_H__
